refactor: consolidate privacy and lpks info in content directory, use consistent implementations
